caf: Add common BLE events
Change adds common BLE events to CAF.

Jira: NCSDK-8532
