+++ applications/nrf_desktop/src/events/Kconfig
@@ -18,22 +18,6 @@ config DESKTOP_INIT_LOG_BATTERY_LEVEL_EVENT
 	bool "Battery level event"
 	default y
 
-config DESKTOP_INIT_LOG_BLE_PEER_EVENT
-	bool "BLE peer event"
-	default y
-
-config DESKTOP_INIT_LOG_BLE_PEER_SEARCH_EVENT
-	bool "BLE peer search event"
-	default y
-
-config DESKTOP_INIT_LOG_BLE_PEER_OPERATION_EVENT
-	bool "BLE peer operation event"
-	default y
-
-config DESKTOP_INIT_LOG_BLE_PEER_CONN_PARAMS_EVENT
-	bool "BLE peer connection parameters event"
-	default y
-
 config DESKTOP_INIT_LOG_BLE_DISC_COMPLETE_EVENT
 	bool "BLE discovery complete event"
 	default y
