+++ applications/nrf_desktop/src/modules/ble_bond.c
@@ -12,7 +12,7 @@
 #define MODULE ble_bond
 #include <caf/events/module_state_event.h>
 #include <caf/events/click_event.h>
-#include "ble_event.h"
+#include <caf/events/ble_common_event.h>
 #include "selector_event.h"
 #include "config_event.h"
 #include <caf/events/power_event.h>
