+++ applications/nrf_desktop/src/modules/ble_conn_params.c
@@ -14,6 +14,7 @@
 
 #define MODULE ble_conn_params
 #include <caf/events/module_state_event.h>
+#include <caf/events/ble_common_event.h>
 #include "ble_event.h"
 
 #include <logging/log.h>
