+++ applications/nrf_desktop/src/modules/ble_discovery.c
@@ -14,6 +14,7 @@
 #define MODULE ble_discovery
 #include <caf/events/module_state_event.h>
 
+#include <caf/events/ble_common_event.h>
 #include "ble_event.h"
 #include "ble_discovery_def.h"
 #include "dev_descr.h"
