+++ applications/nrf_desktop/src/modules/ble_latency.c
@@ -7,6 +7,7 @@
 #include <zephyr/types.h>
 #include <bluetooth/conn.h>
 
+#include <caf/events/ble_common_event.h>
 #include "ble_event.h"
 #include "config_event.h"
 #include <caf/events/power_event.h>
