+++ applications/nrf_desktop/src/modules/ble_scan.c
@@ -15,7 +15,7 @@
 #define MODULE ble_scan
 #include <caf/events/module_state_event.h>
 #include "hid_event.h"
-#include "ble_event.h"
+#include <caf/events/ble_common_event.h>
 
 #include "ble_scan_def.h"
