+++ applications/nrf_desktop/src/modules/hid_state.c
@@ -22,7 +22,7 @@
 #include "motion_event.h"
 #include "wheel_event.h"
 #include "hid_event.h"
-#include "ble_event.h"
+#include <caf/events/ble_common_event.h>
 #include "usb_event.h"
 
 #include "hid_keymap.h"
