+++ applications/nrf_desktop/src/modules/hids.c
@@ -16,7 +16,7 @@
 
 #include "hids_event.h"
 #include "hid_event.h"
-#include "ble_event.h"
+#include <caf/events/ble_common_event.h>
 #include "config_event.h"
 
 #include "hid_report_desc.h"
