+++ applications/nrf_desktop/src/modules/smp.c
@@ -12,6 +12,7 @@
 
 #define MODULE smp
 #include <caf/events/module_state_event.h>
+#include <caf/events/ble_common_event.h>"
 #include "ble_event.h"
 #include "dfu_lock.h"
